+++ app/features/chatbox/chatboxContentTemplate.js
@@ -105,7 +105,7 @@ define(['app/constant', 'shared/data'], (constant, GLOBAL) => {
 
         video: `<video width="400" controls><source src="{src}" type="video/mp4">Your browser does not support HTML video.</video>`,
         mess: `
-            <li class="js_li_list_mess {who} {classLocal} messages__item {isFile}" data-id-local="{idLocal}" ${constant.ATTRIBUTE_MESSAGE_ID}="{id}" data-chat-type="{chatType}">
+            <li class="js_li_list_mess {who} {classLocal} messages__item {isFile} {bookmark}" data-id-local="{idLocal}" ${constant.ATTRIBUTE_MESSAGE_ID}="{id}" data-chat-type="{chatType}">
                 <div class="--content">
                     <div class="--heading">
                         <img ${constant.ATTRIBUTE_CHANGE_IMAGE}="{userId}" class="--img avatar" src="{src}" onerror="this.src='/assets/images/user.jpg'">
@@ -134,9 +134,15 @@ define(['app/constant', 'shared/data'], (constant, GLOBAL) => {
                         <div class="xm icon-bars --bars --menu p-cur btn-message-settings {hide_when_removed}"></div>
                     </div>
 
-                    <!-- <div class="message-bookmark-icon">
+                    <div class="message-bookmark-icon">
                         <i class="icon-bookmarks"></i>
-                    </div> -->
+                    </div>
+                </div>
+
+                <div class="show_origin_mess hidden">
+                    <button type="button" class="show_origin_btn btn btn-secondary" data-toggle="tooltip" data-placement="top" title="Show message" sequence_number="{messSequence}">
+                        <i class="icon-comment-o"></i>
+                    </button>
                 </div>
             </li>
         `